requirements.txt
imutils
numpy
opencv-python
matplotlib
pytesseract
Pillow